requirements.txt
flask==3.1.2
request==2.16.0
firebase-admin==7.1.0
supabase==2.22.0
python-dotenv==1.1.1